requirements.txt
python-telegram-bot
pymongo[srv]
dnspython